// src/app/shared/dynamic-form/constants/InputTypes.enum.ts
export enum InputTypes {
    text = "text",
    checkbox = "checkbox",
    color = "color",
    date = "date",
    datetimeLocal = "datetime-local",
    email = "email",
    file = "file",
    number = "number",
    password = "password",
    radio = "radio",
    range = "range",
    tel = "tel",
    time = "time",
    wcb= "wcb_no",
    claim_no = "claim_no"
}