// src/app/shared/dynamic-form/models/configs.class.ts
export class Configs {
    mask?: string;
    icon?: string;
    formControlName?: string;
    min?: any;
    max?: any;
    name?: string;
    readonly?: boolean;
    button_classes?: any[];
    disabled?: boolean;
    skip_validation?: boolean;
    custom_time?: boolean;
    title_case?: boolean;
	add_tag?:boolean;
    dropdownSearch?: boolean;
    maxLength?:any;
    minLength?:any;
    isNumberOnly?:boolean = false;
}